requirements.txt
opencv-python
opencv-contrib-python
pillow
numpy
